scraper/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mp-scraper"
version = "0.1.0"
description = "Scraper de documentos de licitaciones de Mercado Público Chile"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40.0",
    "pymupdf>=1.23.0",  # fitz para PDFs
    "python-docx>=1.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
mp-scraper = "scraper.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/scraper"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
